+++ psap/__init__.py
@@ -1,6 +1,6 @@
 """Top-level package for psap."""
 from .util import export_matrix
-from .psap import run_model
+from .classifier import run_model
 
 __all__ = ["export_matrix", "run_model"]
