+++ psap/util.py
@@ -1,4 +1,4 @@
-from .psap import MakeMatrix
+from .matrix import MakeMatrix
 import datetime
 from pathlib import Path
